refactor(novu): share agent spec validation and split connect UI
Centralize managed agent limits/validation in @novu/shared and decompose the connect TUI preview flow for easier review and maintenance.
